refactor: 불필요한 파일 및 코드 삭제 & 구조 수정 & 버그 수정
