+++ src/app/(home)/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import InitialChat from "./components/InitialChat";
-import MessageInput from "./components/MessageInput2";
+import MessageInput from "./components/MessageInput";
 import Widget from "./components/Widget";
 import Chat from "./chat/[chatId]/components/Chat";
 import useInitialDataStore from "@/store/initialDataStore";
